+++ src/Contracts/KeyLoader.php
@@ -7,4 +7,9 @@ namespace CodeLieutenant\LaravelCrypto\Contracts;
 interface KeyLoader
 {
     public function getKey(): string|array;
+
+    /**
+     * @return array<int, string>
+     */
+    public function getPreviousKeys(): array;
 }