+++ tests/Feature/ServiceProviderLoadingTest.php
@@ -8,8 +8,13 @@ use Illuminate\Encryption\Encrypter;
 use Illuminate\Support\Facades\Config;
 
 test('encrypter resolver', function (string $cipher, string $instance): void {
-    if (str_starts_with($cipher, 'Sodium_') && !LibEncrypter::supported(str_repeat('0', Encryption::tryFrom($cipher)?->keySize() ?? 32), $cipher)) {
-        $this->markTestSkipped("Cipher $cipher is not supported in this environment.");
+    if (str_starts_with($cipher, 'Sodium_')) {
+        $encryption = Encryption::tryFrom($cipher);
+        $keySize = $encryption?->keySize() ?? 32;
+
+        if (! LibEncrypter::supported(str_repeat('0', $keySize), $cipher)) {
+            $this->markTestSkipped("Cipher $cipher is not supported in this environment.");
+        }
     }
 
     Config::set('app.cipher', $cipher);