+++ tests/InMemoryAppKeyKeyLoader.php
@@ -9,12 +9,25 @@ use Illuminate\Support\Str;
 
 class InMemoryAppKeyKeyLoader implements KeyLoader
 {
-    public function __construct(private readonly string $key) {}
+    /**
+     * @param  string[]  $previousKeys
+     */
+    public function __construct(private readonly string $key, private readonly array $previousKeys = []) {}
 
     public function getKey(): string|array
     {
-        return Str::of($this->key)
-            ->remove('base64:', $this->key)
+        return $this->parse($this->key);
+    }
+
+    public function getPreviousKeys(): array
+    {
+        return array_map($this->parse(...), $this->previousKeys);
+    }
+
+    private function parse(string $key): string
+    {
+        return Str::of($key)
+            ->remove('base64:', $key)
             ->fromBase64()
             ->toString();
     }